[HACK] cpp: make it easier to find macro sources
This is a dubiously useful hack, but invaluable when debugging the
preprocessor.

If CPP_DEBUG_DEFINITIONS is in the environment, output to stderr at
every macro definition its name, and the defining file and line number,
such that the operative definition of common macros (such as __P) can be
determined.